feat: nfcg erros, handle creation in ConversationCreationController
