+++ wire-ios/Wire-iOS/Sources/UserInterface/ConversationList/Container/ViewModel/ConversationListViewControllerViewModel+StartUIDelegate.swift
@@ -37,47 +37,13 @@ extension ConversationListViewController.ViewModel: StartUIDelegate {
         }
     }
 
-    func startUI(
-        _ startUI: StartUIViewController,
-        createConversationWith users: UserSet,
-        name: String,
-        allowGuests: Bool,
-        allowServices: Bool,
-        enableReceipts: Bool,
-        encryptionProtocol: EncryptionProtocol,
-        onCompletion: @escaping (_ postCompletionAction: @escaping () -> Void) -> Void
-    ) {
-        guard let viewController = viewController as? UIViewController else { return }
-        let initialized = conversationCreationCoordinator.initialize { [weak self] result in
-            onCompletion {
-                switch result {
-                case .success(let conversation):
-                    viewController.dismissIfNeeded {
-                        delay(0.3) {
-                            ZClientViewController.shared?.select(
-                                conversation: conversation,
-                                focusOnView: true,
-                                animated: true
-                            )
-                        }
-                    }
-                case .failure:
-                    viewController.dismissIfNeeded()
-                }
-                self?.conversationCreationCoordinator.finalize()
-            }
-        }
-        guard initialized else { return }
-        let creatingConversation = conversationCreationCoordinator.createConversation(
-            withUsers: users,
-            name: name,
-            allowGuests: allowGuests,
-            allowServices: allowServices,
-            enableReceipts: enableReceipts,
-            encryptionProtocol: encryptionProtocol)
-        guard creatingConversation else {
-            conversationCreationCoordinator.finalize()
-            return
+    func startUI(_ startUI: StartUIViewController, didCreateConversation conversation: ZMConversation) {
+        startUI.dismissIfNeeded(animated: true) {
+            ZClientViewController.shared?.select(
+                conversation: conversation,
+                focusOnView: true,
+                animated: true
+            )
         }
     }
 
@@ -100,214 +66,5 @@ extension ConversationListViewController.ViewModel: StartUIDelegate {
                 }
             })
         }
-    }
-}
-
-enum GroupConversationCreationResult {
-    enum FailureType {
-        case missingLegalHoldConsent
-        case nonFederatingBackends
-        case other
-    }
-    case success(conversation: ZMConversation)
-    case failure(failureType: FailureType)
-}
-
-class GroupConversationCreationCoordinator {
-    private enum State {
-        case ready
-        case initializing
-        case initialized(completionHandler: (GroupConversationCreationResult) -> Void)
-        case beginingToCreateConversation(completionHandler: (GroupConversationCreationResult) -> Void)
-        case creatingConversation(conversation: ZMConversation, completionHandler: (GroupConversationCreationResult) -> Void)
-        case finalizing
-    }
-    private var state: State = .ready
-
-    private var conversation: ZMConversation? {
-        guard case .creatingConversation(let conversation, _) = state else { return nil }
-        return conversation
-    }
-
-    private var completionHandler: ((GroupConversationCreationResult) -> Void)? {
-        switch state {
-        case .creatingConversation(_, let completionHandler), .initialized(let completionHandler), .beginingToCreateConversation(let completionHandler):
-            return completionHandler
-        case .ready, .finalizing, .initializing:
-            return nil
-        }
-    }
-
-    func initialize(completionHandler: @escaping (GroupConversationCreationResult) -> Void) -> Bool {
-        guard case .ready = state else { return false }
-        NotificationCenter.default.addObserver(self, selector: #selector(missingLegalConsentErrorHandler), name: ZMConversation.missingLegalHoldConsentNotificationName, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(nonFederatingBackendsErrorHandler), name: ZMConversation.nonFederatingBackendsNotificationName, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(unsupportedErrorsHandler), name: ZMConversation.unknownResponseErrorNotificationName, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(conversationCreatedHandler), name: ZMConversation.insertedConversationUpdatedNotificationName, object: nil)
-        state = .initialized(completionHandler: completionHandler)
-        return true
-    }
-
-    func createConversation(
-        withUsers users: UserSet?,
-        name: String?,
-        allowGuests: Bool,
-        allowServices: Bool,
-        enableReceipts: Bool,
-        encryptionProtocol: EncryptionProtocol
-     ) -> Bool {
-         guard case .initialized(let completionHandler) = state else { return false }
-         guard
-             let users = users,
-             let userSession = ZMUserSession.shared()
-         else {
-             return false
-         }
-         state = .beginingToCreateConversation(completionHandler: completionHandler)
-
-         var conversation: ZMConversation?
-
-         userSession.enqueue {
-             conversation = ZMConversation.insertGroupConversation(
-                 moc: userSession.viewContext,
-                 participants: users.materialize(in: userSession.viewContext),
-                 name: name,
-                 team: ZMUser.selfUser().team,
-                 allowGuests: allowGuests,
-                 allowServices: allowServices,
-                 readReceipts: enableReceipts,
-                 messageProtocol: encryptionProtocol == .mls ? .mls : .proteus
-             )
-         } completionHandler: { [weak self] in
-             guard let self = self, let conversation = conversation else { return }
-             self.state = .creatingConversation(conversation: conversation, completionHandler: completionHandler)
-         }
-
-         return true
-    }
-
-    func createConversation(
-        withParticipants users: UserSet,
-        name: String?
-     ) -> Bool {
-         guard case .initialized(let completionHandler) = state else { return false }
-         guard let userSession = ZMUserSession.shared() else { return false }
-         state = .beginingToCreateConversation(completionHandler: completionHandler)
-
-         var conversation: ZMConversation?
-
-         userSession.enqueue {
-             conversation = ZMConversation.insertGroupConversation(session: userSession,
-                                                                   participants: Array(users),
-                                                                   name: name,
-                                                                   team: ZMUser.selfUser().team)
-         } completionHandler: { [weak self] in
-             guard let self = self, let conversation = conversation else { return }
-             self.state = .creatingConversation(conversation: conversation, completionHandler: completionHandler)
-         }
-
-         return true
-    }
-
-    func finalize() {
-        state = .finalizing
-        NotificationCenter.default.removeObserver(self, name: ZMConversation.missingLegalHoldConsentNotificationName, object: nil)
-        NotificationCenter.default.removeObserver(self, name: ZMConversation.nonFederatingBackendsNotificationName, object: nil)
-        NotificationCenter.default.removeObserver(self, name: ZMConversation.unknownResponseErrorNotificationName, object: nil)
-        NotificationCenter.default.removeObserver(self, name: ZMConversation.insertedConversationUpdatedNotificationName, object: nil)
-        state = .ready
-    }
-
-    @objc func conversationCreatedHandler(notification: Notification) {
-        guard let conversation = extractConversation(notification: notification)
-            else { return }
-        completionHandler?(.success(conversation: conversation))
-    }
-
-    @objc func unsupportedErrorsHandler(notification: Notification) {
-        guard extractConversation(notification: notification) != nil
-            else { return }
-        completionHandler?(.failure(failureType: .other))
-    }
-
-    @objc func missingLegalConsentErrorHandler(notification: Notification) {
-        guard extractConversation(notification: notification) != nil
-            else { return }
-        typealias ConversationError = L10n.Localizable.Error.Conversation
-        UIAlertController.showErrorAlert(
-            title: ConversationError.title,
-            message: ConversationError.missingLegalholdConsent,
-            completion: { [weak self] _ in
-                self?.completionHandler?(.failure(failureType: .missingLegalHoldConsent))
-            }
-        )
-    }
-
-    @objc func nonFederatingBackendsErrorHandler(notification: Notification) {
-            guard extractConversation(notification: notification) != nil
-                else { return }
-        guard let userInfo = notification.userInfo,
-              userInfo.keys.contains(ZMConversation.UserInfoKeys.nonFederatingBackends.rawValue),
-              let nonFederatingBackends = userInfo[ZMConversation.UserInfoKeys.nonFederatingBackends.rawValue] as? NonFederatingBackendsTuple
-            else { return }
-        let alert = UIAlertController.notFullyConnectedGraphAlert(
-            backends: nonFederatingBackends,
-            completion: { [weak self] action in
-                switch action {
-                case .editParticipantsList:
-                    self?.finalize()
-                case .discardGroupCreation:
-                    self?.completionHandler?(.failure(failureType: .nonFederatingBackends))
-                case .learnMore:
-                    self?.finalize()
-                }
-            }
-        )
-        alert.presentTopmost()
-    }
-
-    private func extractConversation(notification: Notification) -> ZMConversation? {
-        guard
-            let userInfo = notification.userInfo,
-            userInfo.keys.contains(NotificationInContext.objectInNotificationKey),
-            let notificationConversation = userInfo[NotificationInContext.objectInNotificationKey] as? ZMConversation,
-            let conversation = self.conversation,
-            conversation.objectID == notificationConversation.objectID
-        else { return nil }
-        return conversation
-    }
-}
-
-enum NonFullyConnectedGraphAction {
-    case editParticipantsList
-    case discardGroupCreation
-    case learnMore
-}
-
-extension UIAlertController {
-    static func notFullyConnectedGraphAlert(backends: NonFederatingBackendsTuple, completion: @escaping (NonFullyConnectedGraphAction) -> Void) -> UIAlertController {
-        let alert = UIAlertController(
-            title: "Group can’t be created",
-            message: "People from backends \(backends.backendA) and \(backends.backendB) can’t join the same group conversation.\nTo create the group, remove affected participants.", preferredStyle: .alert)
-        let editParticipantsAction = UIAlertAction(
-            title: "Edit Participants List",
-            style: .default,
-            handler: { _ in completion(.editParticipantsList) }
-        )
-        let discardGroupCreationAction = UIAlertAction(
-            title: "Discard Group Creation",
-            style: .default,
-            handler: { _ in completion(.discardGroupCreation) }
-        )
-        let learnMoreAction = UIAlertAction(
-            title: "Learn More",
-            style: .default,
-            handler: { _ in completion(.learnMore) }
-        )
-        alert.addAction(editParticipantsAction)
-        alert.addAction(discardGroupCreationAction)
-        alert.addAction(learnMoreAction)
-        alert.preferredAction = editParticipantsAction
-        return alert
     }
 }
